pyproject.toml
[project]
name = "lock-detech"
version = "0.1.0"
description = "智能锁检测系统 - 基于YOLOv8的锁状态识别API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "pandas>=2.3.2",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "ultralytics>=8.3.192",
    "uvicorn>=0.35.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0.0",
    "scikit-image>=0.19.0",
    "opencv-python>=4.11.0.86",
    "numpy>=2.3.2",
    "httpx>=0.28.1",
]
